+++ packages/riverpod/lib/src/providers/future_provider.dart
@@ -14,6 +14,14 @@ import 'stream_provider.dart' show StreamProvider;
 @internal
 base mixin $FutureProvider<StateT, RefT> on ProviderBase<AsyncValue<StateT>> {
   FutureOr<StateT> create(RefT ref);
+
+  // TODO doc
+  Override overrideWithValue(AsyncValue<StateT> value) {
+    return $ProviderOverride(
+      origin: this,
+      providerOverride: $AsyncValueProvider<StateT>(value),
+    );
+  }
 }
 
 /// {@template riverpod.future_provider}