+++ packages/riverpod/test/old/legacy_providers/future_provider/auto_dispose_family_future_provider_test.dart
@@ -13,7 +13,7 @@ void main() {
     expect(provider(0).argument, 0);
   });
 
-  group('scoping an override overrides all the associated subproviders', () {
+  group('scoping an override overrides all the associated sub-providers', () {
     test('when passing the provider itself', () async {
       final provider = FutureProvider.autoDispose.family<int, int>(
         (ref, _) async => 0,