+++ packages/riverpod/test/old/legacy_providers/future_provider/future_provider_test.dart
@@ -388,7 +388,7 @@ void main() {
     verifyNoMoreInteractions(listener);
   });
 
-  group('scoping an override overrides all the associated subproviders', () {
+  group('scoping an override overrides all the associated sub-providers', () {
     test('when passing the provider itself', () async {
       final provider = FutureProvider(
         (ref) async => 0,
@@ -409,22 +409,27 @@ void main() {
       ]);
     });
 
-    // test('when using provider.overrideWithValue', () async {
-    //   final provider = FutureProvider((ref) async => 0);
-    //   final root = ProviderContainer.test();
-    //   final container = ProviderContainer.test(parent: root, overrides: [
-    //     provider.overrideWithValue(const AsyncValue.data(42)),
-    //   ]);
+    test('when using provider.overrideWithValue', () async {
+      final provider = FutureProvider(
+        (ref) async => 0,
+        dependencies: const [],
+      );
+      final root = ProviderContainer.test();
+      final container = ProviderContainer.test(
+        parent: root,
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.data(42)),
+        ],
+      );
 
-    //   expect(await container.read(provider.future), 42);
-    //   expect(container.read(provider), const AsyncValue.data(42));
-    //   expect(root.getAllProviderElementsInOrder(), isEmpty);
-    //   expect(container.getAllProviderElementsInOrder(), [
-    //     isA<ProviderElementBase<Object?>>().having((e) => e.origin, 'origin', provider),
-    //     isA<ProviderElementBase<Object?>>()
-    //         .having((e) => e.origin, 'origin', provider.future)
-    //   ]);
-    // });
+      expect(await container.read(provider.future), 42);
+      expect(container.read(provider), const AsyncValue.data(42));
+      expect(root.getAllProviderElementsInOrder(), isEmpty);
+      expect(container.getAllProviderElementsInOrder(), [
+        isA<ProviderElementBase<Object?>>()
+            .having((e) => e.origin, 'origin', provider),
+      ]);
+    });
 
     test('when using provider.overrideWith', () async {
       final provider = FutureProvider(
@@ -449,24 +454,26 @@ void main() {
     });
   });
 
-  // test(
-  //     'when overridden with an error but provider.future is not listened to, it should not emit an error to the zone',
-  //     () async {
-  //   final error = Error();
-  //   final future = FutureProvider<int>((ref) async => 0);
+  test(
+      'when overridden with an error but provider.future is not listened to, '
+      'it should not emit an error to the zone', () async {
+    final error = Error();
+    final future = FutureProvider<int>((ref) async => 0);
 
-  //   final container = ProviderContainer.test(overrides: [
-  //     future.overrideWithValue(AsyncValue.error(error)),
-  //   ]);
-  //   addTearDown(container.dispose);
+    final container = ProviderContainer.test(
+      overrides: [
+        future.overrideWithValue(AsyncValue.error(error, StackTrace.empty)),
+      ],
+    );
+    addTearDown(container.dispose);
 
-  //   expect(
-  //     container.read(future),
-  //     AsyncValue<int>.error(error),
-  //   );
+    expect(
+      container.read(future),
+      AsyncValue<int>.error(error, StackTrace.empty),
+    );
 
-  //   // the test will naturally fail if a non-caught future is created
-  // });
+    // the test will naturally fail if a non-caught future is created
+  });
 
   test('throwing inside "create" result in an AsyncValue.error', () {
     // ignore: only_throw_errors
@@ -701,285 +708,301 @@ void main() {
     verifyNoMoreInteractions(listener);
   });
 
-  // group('mock as value', () {
-  //   test('value immediately then other value', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
-
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       completion(42),
-  //     );
-
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), const AsyncValue.data(42));
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.data(21)),
-  //     ]);
+  group('mock as value', () {
+    test('value immediately then other value', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.data(42)),
+        ],
+      );
+
+      await expectLater(
+        container.read(provider.future),
+        completion(42),
+      );
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), const AsyncValue.data(42));
+
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.data(21)),
+      ]);
+
+      await expectLater(
+        container.read(provider.future),
+        completion(21),
+      );
+      expect(sub.read(), const AsyncValue.data(21));
+    });
+
+    test('value immediately then error', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.data(42)),
+        ],
+      );
+
+      await expectLater(
+        container.read(provider.future),
+        completion(42),
+      );
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), const AsyncValue.data(42));
+
+      container.updateOverrides([
+        provider
+            .overrideWithValue(const AsyncValue.error(21, StackTrace.empty)),
+      ]);
+
+      await expectLater(
+        container.read(provider.future),
+        throwsA(21),
+      );
+      expect(
+        sub.read(),
+        const AsyncValue<int>.error(21, StackTrace.empty)
+            .copyWithPrevious(const AsyncData(42)),
+      );
+    });
+
+    test('value immediately then loading', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.data(42)),
+        ],
+      );
+
+      final future = container.read(provider.future);
+
+      await expectLater(
+        future,
+        completion(42),
+      );
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), const AsyncValue<int>.data(42));
+
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.loading()),
+      ]);
+
+      expect(container.read(provider.future), isNot(future));
+      expect(
+        sub.read(),
+        const AsyncLoading<int>()
+            .copyWithPrevious(const AsyncValue<int>.data(42)),
+      );
+    });
+
+    test('loading immediately then value', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.loading()),
+        ],
+      );
+
+      final future = container.read(provider.future);
 
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       completion(21),
-  //     );
-  //     expect(sub.read(), const AsyncValue.data(21));
-  //   });
+      final sub = container.listen(provider, (_, __) {});
 
-  //   test('value immediately then error', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
+      expect(sub.read(), const AsyncValue<int>.loading());
 
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       completion(42),
-  //     );
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.data(42)),
+      ]);
 
-  //     final sub = container.listen(provider, (_, __) {});
+      expect(sub.read(), const AsyncValue<int>.data(42));
 
-  //     expect(sub.read(), const AsyncValue.data(42));
+      await expectLater(future, completion(42));
+    });
 
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.error(21)),
-  //     ]);
+    test('loading immediately then error', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.loading()),
+        ],
+      );
 
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       throwsA(21),
-  //     );
-  //     expect(
-  //       sub.read(),
-  //       const AsyncValue<int>.error(21).copyWithPrevious(const AsyncData(42)),
-  //     );
-  //   });
+      final future = container.read(provider.future);
 
-  //   test('value immediately then loading', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
+      final sub = container.listen(provider, (_, __) {});
 
-  //     final future = container.read(provider.future);
+      expect(sub.read(), const AsyncValue<int>.loading());
 
-  //     await expectLater(
-  //       future,
-  //       completion(42),
-  //     );
+      final stackTrace = StackTrace.current;
 
-  //     final sub = container.listen(provider, (_, __) {});
+      container.updateOverrides([
+        provider.overrideWithValue(AsyncValue.error(42, stackTrace)),
+      ]);
 
-  //     expect(sub.read(), const AsyncValue<int>.data(42));
+      expect(sub.read(), AsyncValue<int>.error(42, stackTrace));
 
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
+      await expectLater(future, throwsA(42));
+    });
 
-  //     expect(container.read(provider.future), isNot(future));
-  //     expect(
-  //       sub.read(),
-  //       const AsyncLoading<int>()
-  //           .copyWithPrevious(const AsyncValue<int>.data(42)),
-  //     );
-  //   });
+    test('loading immediately then loading', () async {
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(const AsyncValue.loading()),
+        ],
+      );
+      final listener = Listener<AsyncValue<int>>();
 
-  //   test('loading immediately then value', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
+      final future = container.read(provider.future);
 
-  //     final future = container.read(provider.future);
+      container.listen(provider, listener.call, fireImmediately: true);
 
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), const AsyncValue<int>.loading());
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
+      verifyOnly(listener, listener(null, const AsyncValue<int>.loading()));
 
-  //     expect(sub.read(), const AsyncValue<int>.data(42));
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.loading()),
+      ]);
 
-  //     await expectLater(future, completion(42));
-  //   });
+      verifyNoMoreInteractions(listener);
 
-  //   test('loading immediately then error', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.data(42)),
+      ]);
 
-  //     final future = container.read(provider.future);
+      verifyOnly(
+        listener,
+        listener(const AsyncValue.loading(), const AsyncValue.data(42)),
+      );
+      await expectLater(future, completion(42));
+    });
 
-  //     final sub = container.listen(provider, (_, __) {});
+    test('error immediately then different error', () async {
+      final stackTrace = StackTrace.current;
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(AsyncValue.error(42, stackTrace)),
+        ],
+      );
+
+      await expectLater(
+        container.read(provider.future),
+        throwsA(42),
+      );
 
-  //     expect(sub.read(), const AsyncValue<int>.loading());
+      final sub = container.listen(provider, (_, __) {});
 
-  //     final stackTrace = StackTrace.current;
+      expect(sub.read(), AsyncValue<int>.error(42, stackTrace));
 
-  //     container.updateOverrides([
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(42, stackTrace: stackTrace)),
-  //     ]);
-
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stackTrace));
-
-  //     await expectLater(future, throwsA(42));
-  //   });
-
-  //   test('loading immediately then loading', () async {
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
-  //     final listener = Listener<AsyncValue<int>>();
-
-  //     final future = container.read(provider.future);
-
-  //     container.listen(provider, listener, fireImmediately: true);
-
-  //     verifyOnly(listener, listener(null, const AsyncValue<int>.loading()));
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
-
-  //     verifyNoMoreInteractions(listener);
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
-
-  //     verifyOnly(
-  //       listener,
-  //       listener(const AsyncValue.loading(), const AsyncValue.data(42)),
-  //     );
-  //     await expectLater(future, completion(42));
-  //   });
-
-  //   test('error immediately then different error', () async {
-  //     final stackTrace = StackTrace.current;
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(42, stackTrace: stackTrace)),
-  //     ]);
-
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       throwsA(42),
-  //     );
-
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stackTrace));
-
-  //     container.updateOverrides([
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(21, stackTrace: stackTrace)),
-  //     ]);
-
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       throwsA(21),
-  //     );
-  //     expect(sub.read(), AsyncValue<int>.error(21, stackTrace: stackTrace));
-  //   });
-
-  //   test('error immediately then different stacktrace', () async {
-  //     final stackTrace = StackTrace.current;
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(42, stackTrace: stackTrace)),
-  //     ]);
-
-  //     final future = container.read(provider.future);
-
-  //     await expectLater(future, throwsA(42));
-
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stackTrace));
-
-  //     final stack2 = StackTrace.current;
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(AsyncValue.error(42, stackTrace: stack2)),
-  //     ]);
-
-  //     expect(
-  //       container.read(provider.future),
-  //       isNot(future),
-  //     );
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       throwsA(42),
-  //     );
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stack2));
-  //   });
-
-  //   test('error immediately then data', () async {
-  //     final stackTrace = StackTrace.current;
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(42, stackTrace: stackTrace)),
-  //     ]);
-
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       throwsA(42),
-  //     );
-
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stackTrace));
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.data(42)),
-  //     ]);
-
-  //     await expectLater(
-  //       container.read(provider.future),
-  //       completion(42),
-  //     );
-  //     expect(
-  //       sub.read(),
-  //       const AsyncValue<int>.data(42)
-  //           .copyWithPrevious(AsyncError(42, stackTrace: stackTrace)),
-  //     );
-  //   });
-
-  //   test('error immediately then loading', () async {
-  //     final stackTrace = StackTrace.current;
-  //     final provider = FutureProvider((_) async => 0);
-  //     final container = ProviderContainer.test(overrides: [
-  //       provider
-  //           .overrideWithValue(AsyncValue.error(42, stackTrace: stackTrace)),
-  //     ]);
-
-  //     final future = container.read(provider.future);
-  //     await expectLater(future, throwsA(42));
-
-  //     final sub = container.listen(provider, (_, __) {});
-
-  //     expect(sub.read(), AsyncValue<int>.error(42, stackTrace: stackTrace));
-
-  //     container.updateOverrides([
-  //       provider.overrideWithValue(const AsyncValue.loading()),
-  //     ]);
-
-  //     expect(container.read(provider.future), isNot(future));
-  //     expect(
-  //       sub.read(),
-  //       const AsyncLoading<int>()
-  //           .copyWithPrevious(AsyncError<int>(42, stackTrace: stackTrace)),
-  //     );
-  //   });
-  // });
+      container.updateOverrides([
+        provider.overrideWithValue(AsyncValue.error(21, stackTrace)),
+      ]);
+
+      await expectLater(
+        container.read(provider.future),
+        throwsA(21),
+      );
+      expect(sub.read(), AsyncValue<int>.error(21, stackTrace));
+    });
+
+    test('error immediately then different stacktrace', () async {
+      final stackTrace = StackTrace.current;
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(AsyncValue.error(42, stackTrace)),
+        ],
+      );
+
+      final future = container.read(provider.future);
+
+      await expectLater(future, throwsA(42));
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), AsyncValue<int>.error(42, stackTrace));
+
+      final stack2 = StackTrace.current;
+
+      container.updateOverrides([
+        provider.overrideWithValue(AsyncValue.error(42, stack2)),
+      ]);
+
+      expect(
+        container.read(provider.future),
+        isNot(future),
+      );
+      await expectLater(
+        container.read(provider.future),
+        throwsA(42),
+      );
+      expect(sub.read(), AsyncValue<int>.error(42, stack2));
+    });
+
+    test('error immediately then data', () async {
+      final stackTrace = StackTrace.current;
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(AsyncValue.error(42, stackTrace)),
+        ],
+      );
+
+      await expectLater(
+        container.read(provider.future),
+        throwsA(42),
+      );
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), AsyncValue<int>.error(42, stackTrace));
+
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.data(42)),
+      ]);
+
+      await expectLater(
+        container.read(provider.future),
+        completion(42),
+      );
+      expect(
+        sub.read(),
+        const AsyncValue<int>.data(42)
+            .copyWithPrevious(AsyncError(42, stackTrace)),
+      );
+    });
+
+    test('error immediately then loading', () async {
+      final stackTrace = StackTrace.current;
+      final provider = FutureProvider((_) async => 0);
+      final container = ProviderContainer.test(
+        overrides: [
+          provider.overrideWithValue(AsyncValue.error(42, stackTrace)),
+        ],
+      );
+
+      final future = container.read(provider.future);
+      await expectLater(future, throwsA(42));
+
+      final sub = container.listen(provider, (_, __) {});
+
+      expect(sub.read(), AsyncValue<int>.error(42, stackTrace));
+
+      container.updateOverrides([
+        provider.overrideWithValue(const AsyncValue.loading()),
+      ]);
+
+      expect(container.read(provider.future), isNot(future));
+      expect(
+        sub.read(),
+        const AsyncLoading<int>()
+            .copyWithPrevious(AsyncError<int>(42, stackTrace)),
+      );
+    });
+  });
 }