+++ packages/riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Breaking** `AsyncValue.value` now returns `null` during errors.
 - **Breaking** removed `AsyncValue.valueOrNull` (use `.value` instead).
+- `Stream/FutureProvider.overrideWithValue` was added back.
 - Updated `AsyncValue` documentations to use pattern matching.
 
 ## 3.0.0-dev.3 - 2023-11-27